Add explicit pubkey verification in Kex test
Added an explicit verification that two public keys differ, just
to make sure that we aren't always generating the same public keys.
This should hopefully never become relevant, but if we for some
reason start generating the same keypair every time, it should raise
some eyebrows, and the test is basically free, computationally.