Replace CursorAdapter with FriendListAdapter
The CursorAdapter required knowledge about internals of the database
to work.  As I want to encapsulate all knowledge about the Database
in the DatabaseService and related helpers, that was unacceptable.
It was replaced with a custom Adapter that takes a List of Friends
to disply.  The relevant changes were also made to the DatabaseService
to support this use, and the Friend class to make it easier to
instantiate a new Friend from the database.