Add warning on failed verification
If a verification has explicitly failed, saving the new contact is
now impossible.  The user can either try scanning again, or start
over.  When starting over, there is a small display bug with the
background animation => You have to click the item once to get it
to be interactive, then click it again to trigger the action.  I'll
have to see if I can find out why that is so.