Add FriendManagement helper, add delete func
Added a static FriendManagement helper class that contains static
functions to perform Friend Management.  That way, a unified API
can be used for all Friend operations.
Also added a deletion function for friends, which will delete the
Friend and related keys (through the ON DELETE CASCADE on the
foreign key definition) from the database.