Add serialization and deserialization to GPSTrack
GPSTrack can now serialize and deserialize to / from protobuf.
Also added a test, but that one should be taken with a grain of
salt, as strictly speaking protobuf "repeated" Lists do not
guarantee that the order will be preserved, which is a prerequisite
for the test to run successfully.  But since Location.equals()
appearently checks pointer equality, there is no better way. :/