Add owner column to database
In preparation for the upcoming changes needed to allow for saving
shared data in the database, this commit adds an "owner"-column to
the database.  This column will either be set to -1, indicating
that the user of the device is the owner, or to the ID of the
friend in the database.  However, it does not use a foreign key,
and thus deletions of friends will not automatically propagate.