Add first version of granularity-aware sharing
Right now, we can share with a different granularity, but the display
will not be very nice if the path data is not shared.  The description
can also be set, but will not be displayed.  Also, if a track has been
shared with one granularity, it cannot be shared with any other, and
any attempt to do so will result in sharing the old granularity.  This
is terrible and should be fixed.