Add share granularity locking
Shares are now locked to the first shared granularity and description.
Also, friends who have already received a share will now be pre-
selected in the list of share recipients (but the selection will not
be "locked in" yet).  Friends that are un-checked after having already
received the share will _not_ loose access to the share (i.e. this is no
revocation UI).  Friends that have already received the share _will_
currently receive the share a second time (I'll fix that next).